requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dataclasses-json==0.6.7
distro==1.9.0
emoji==2.14.1
eval_type_backport==0.2.2
faiss-cpu==1.10.0
filetype==1.2.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.3.28
lxml==5.3.2
marshmallow==3.26.1
multidict==6.4.2
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.4
olefile==0.47
openai==1.72.0
orjson==3.10.16
packaging==24.2
propcache==0.3.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
unstructured==0.17.2
unstructured-client==0.32.3
urllib3==2.3.0
webencodings==0.5.1
wrapt==1.17.2
yarl==1.19.0
zstandard==0.23.0
